requirements.txt
ffmpeg-python==0.2.0
Flask==2.2.2
flask_cors==3.0.10
youtube_transcript_api==0.4.1
openai==1.12.0
openai-whisper==20231117
python-dotenv==0.19.0
Werkzeug==2.2.2